Fixed deprecated issues (macOS 10.14+)
Fixed issues with Xcode 11.2 and macOS 10.15 Catalina deprecating the use of NSVisualEffectView.Material.appearanceBased. This allowed for seemless matching of the blurred background with whatever content layed beneath the window – for all versions of macOS 10.10-10.14.

As of 10.14, Apple has introduced NSVisualEffectView.underWindowBackground, re-enabling this functionality. However, this new setup is only compatible with macOS 10.14 and higher.

WARNING (macOS 10.3 and lower):
If your project is currently being provided to users with 10.13 or lower, this new implementation will not work. You will need to use the build prior to this one (reference below).

https://github.com/revblaze/WKBlurEffect/commit/947bb5b0d4d50f3d20005069e9a88ac0b23f257c